+++ join/join-testnet.sh
@@ -60,8 +60,18 @@ if [ -z "$NETWORK_PRESENT" ]; then
 fi
 
 if $RESET_CHAIN; then
-  rm -rf "$ROOT_DIRECTORY"
+  echo
+  echo "WARNING! This will erase all previously stored data. Your node will catch up from the beginning"
+  printf "Do you wish to proceed \"y/n\" ?  "
+  read -r REPLY
+  if [ $REPLY = "y" ]; then
+    echo "Resetting state"
+    rm -rf "$ROOT_DIRECTORY"
+  else
+    echo "Proceeding without resetting state"
+  fi
 fi
+
 
 mkdir -p "$ROOT_DIRECTORY"
 
@@ -72,11 +82,15 @@ CORE_DIRECTORY="${ROOT_DIRECTORY}/.core"
 mkdir -p "$CORE_DIRECTORY"
 
 if [ ! -f "${SHARED_DIRECTORY}/genesis.json" ]; then
-  curl https://axelar-testnet.s3.us-east-2.amazonaws.com/genesis.json -o "${SHARED_DIRECTORY}/genesis.json"
+  curl -s https://axelar-testnet.s3.us-east-2.amazonaws.com/genesis.json -o "${SHARED_DIRECTORY}/genesis.json"
 fi
 
 echo "Downloading latest persistent peers"
-curl https://axelar-testnet.s3.us-east-2.amazonaws.com/persistent-peers.txt -o "${SHARED_DIRECTORY}/persistent-peers.txt"
+curl -s https://axelar-testnet.s3.us-east-2.amazonaws.com/persistent-peers.txt -o "${SHARED_DIRECTORY}/persistent-peers.txt"
+
+if [ ! -f "${SHARED_DIRECTORY}/seeds.txt" ]; then
+  curl https://axelar-testnet.s3.us-east-2.amazonaws.com/seeds.txt -o "${SHARED_DIRECTORY}/seeds.txt"
+fi
 
 echo "Overwriting stale config.toml to config directory"
 cp "${GIT_ROOT}/join/config.toml" "${SHARED_DIRECTORY}/config.toml"
@@ -107,6 +121,7 @@ docker run                                             \
   --env CONFIG_PATH=/root/shared/                      \
   --env AXELAR_MNEMONIC_PATH=$AXELAR_MNEMONIC_PATH     \
   --env TENDERMINT_KEY_PATH=$TENDERMINT_KEY_PATH       \
+  --env PEERS_FILE=/root/shared/seeds.txt              \
   -v "${CORE_DIRECTORY}/:/root/.axelar"                \
   -v "${SHARED_DIRECTORY}:/root/shared"                \
   "axelarnet/axelar-core:${AXELAR_CORE_VERSION}" startNodeProc
